autogpt_platform/backend/pyproject.toml
[tool.poetry]
name = "autogpt-platform-backend"
version = "0.4.9"
description = "A platform for building AI-powered agentic workflows"
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"
packages = [{ include = "backend", format = "sdist" }]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aio-pika = "^9.5.5"
anthropic = "^0.49.0"
apscheduler = "^3.11.0"
autogpt-libs = { path = "../autogpt_libs", develop = true }
bleach = { extras = ["css"], version = "^6.2.0" }
click = "^8.1.7"
cryptography = "^43.0"
discord-py = "^2.5.2"
e2b-code-interpreter = "^1.1.1"
fastapi = "^0.115.12"
feedparser = "^6.0.11"
flake8 = "^7.2.0"
google-api-python-client = "^2.166.0"
google-auth-oauthlib = "^1.2.1"
google-cloud-storage = "^3.1.0"
googlemaps = "^4.10.0"
gravitasml = "^0.1.3"
groq = "^0.20.0"
jinja2 = "^3.1.6"
jsonref = "^1.1.0"
jsonschema = "^4.22.0"
launchdarkly-server-sdk = "^9.10.0"
mem0ai = "^0.1.80"
moviepy = "^2.1.2"
ollama = "^0.4.1"
openai = "^1.70.0"
pika = "^1.3.2"
pinecone = "^5.3.1"
poetry = "^2.1.2"
postmarker = "^1.0"
praw = "~7.8.1"
prisma = "^0.15.0"
psutil = "^7.0.0"
psycopg2-binary = "^2.9.10"
pydantic = { extras = ["email"], version = "^2.11.1" }
pydantic-settings = "^2.8.1"
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
python-dotenv = "^1.1.0"
python-multipart = "^0.0.20"
redis = "^5.2.0"
replicate = "^1.0.4"
sentry-sdk = {extras = ["anthropic", "fastapi", "launchdarkly", "openai", "sqlalchemy"], version = "^2.25.1"}
sqlalchemy = "^2.0.40"
strenum = "^0.4.9"
stripe = "^11.5.0"
supabase = "2.15.0"
tenacity = "^9.0.0"
todoist-api-python = "^2.1.7"
tweepy = "^4.14.0"
uvicorn = { extras = ["standard"], version = "^0.34.0" }
websockets = "^14.2"
youtube-transcript-api = "^0.6.2"
zerobouncesdk = "^1.1.1"
# NOTE: please insert new dependencies in their alphabetical location
prometheus-client = "^0.21.1"
tiktokapi = "^7.1.0"
yt-dlp = "^2025.4.30"
pydub = "^0.25.1"
scipy = "^1.15.3"
torch = "^2.7.0"
transformers = "^4.52.0"
openai-whisper = "^20240930"
suno-bark = {git = "https://github.com/suno-ai/bark.git"}
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
aiohappyeyeballs = "^2.6.1"
black = "^24.10.0"
faker = "^33.3.1"
httpx = "^0.28.1"
isort = "^5.13.2"
poethepoet = "^0.33.1"
pyright = "^1.1.398"
pytest-mock = "^3.14.0"
pytest-watcher = "^0.4.2"
requests = "^2.32.3"
ruff = "^0.11.2"
# NOTE: please insert new dependencies in their alphabetical location

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "backend.app:main"
rest = "backend.rest:main"
ws = "backend.ws:main"
scheduler = "backend.scheduler:main"
executor = "backend.exec:main"
cli = "backend.cli:main"
format = "linter:format"
lint = "linter:lint"
test = "run_tests:test"

[tool.isort]
profile = "black"

[tool.pytest-watcher]
now = false
clear = true
delay = 0.2
runner = "pytest"
runner_args = []
patterns = ["*.py"]
ignore_patterns = []

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
